Move shannonComponent in a separate class
